+++ wp_api/src/wp_com/endpoint/stats_insights_endpoint.rs
@@ -1,15 +1,12 @@
 use crate::{
     request::endpoint::{AsNamespace, DerivedRequest},
-    wp_com::{
-        WpComNamespace, WpComSiteId,
-        stats_insights::{StatsInsightsParams, StatsInsightsResponse},
-    },
+    wp_com::{WpComNamespace, WpComSiteId, stats_insights::StatsInsightsResponse},
 };
 use wp_derive_request_builder::WpDerivedRequest;
 
 #[derive(WpDerivedRequest)]
 enum StatsInsightsRequest {
-    #[get(url = "/sites/<wp_com_site_id>/stats/insights", params = &StatsInsightsParams, output = StatsInsightsResponse)]
+    #[get(url = "/sites/<wp_com_site_id>/stats/insights", output = StatsInsightsResponse)]
     GetStatsInsights,
 }