+++ wp_api/src/wp_com/stats_city_views.rs
@@ -2,7 +2,6 @@ use crate::{
     date::WpDateString,
     impl_as_query_value_from_to_string,
     url_query::{AppendUrlQueryPairs, QueryPairs, QueryPairsExtension},
-    wp_com::language::WPComLanguage,
 };
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -57,9 +56,6 @@ pub struct StatsCityViewsParams {
     /// The number of days to include in the response.
     #[uniffi(default = None)]
     pub days: Option<u32>,
-    /// The locale for the response.
-    #[uniffi(default = None)]
-    pub locale: Option<WPComLanguage>,
     /// Whether to return a summary of the data.
     ///
     /// - `true` (default): Response contains `summary` field with aggregated city
@@ -80,7 +76,6 @@ impl Default for StatsCityViewsParams {
             max: None,
             num: None,
             days: None,
-            locale: None,
             summarize: true,
         }
     }
@@ -95,7 +90,6 @@ impl AppendUrlQueryPairs for StatsCityViewsParams {
             .append_option_query_value_pair("max", self.max.as_ref())
             .append_option_query_value_pair("num", self.num.as_ref())
             .append_option_query_value_pair("days", self.days.as_ref())
-            .append_option_query_value_pair("locale", self.locale.as_ref())
             .append_query_value_pair("summarize", &(self.summarize as u32));
     }
 }
@@ -198,7 +192,6 @@ mod tests {
             max: Some(10),
             num: Some(1),
             days: Some(1),
-            locale: Some(WPComLanguage::English),
             summarize: true,
         };
 
@@ -207,7 +200,7 @@ mod tests {
 
         assert_eq!(
             query_pairs.finish().as_str(),
-            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/location-views/city?period=day&date=2026-02-05&start_date=2026-01-30&max=10&num=1&days=1&locale=en&summarize=1"
+            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/location-views/city?period=day&date=2026-02-05&start_date=2026-01-30&max=10&num=1&days=1&summarize=1"
         );
     }
 
@@ -222,7 +215,6 @@ mod tests {
             period: Some(StatsCityViewsPeriod::Day),
             date: Some(WpDateString::new("2026-02-05".to_string())),
             start_date: Some(WpDateString::new("2026-01-30".to_string())),
-            locale: Some(WPComLanguage::English),
             ..Default::default()
         };
 
@@ -231,7 +223,7 @@ mod tests {
 
         assert_eq!(
             query_pairs.finish().as_str(),
-            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/location-views/city?period=day&date=2026-02-05&start_date=2026-01-30&locale=en&summarize=1"
+            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/location-views/city?period=day&date=2026-02-05&start_date=2026-01-30&summarize=1"
         );
     }