+++ wp_api/src/wp_com/stats_clicks.rs
@@ -2,7 +2,6 @@ use crate::{
     date::WpDateString,
     impl_as_query_value_from_to_string,
     url_query::{AppendUrlQueryPairs, QueryPairs, QueryPairsExtension},
-    wp_com::language::WPComLanguage,
 };
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -53,9 +52,6 @@ pub struct StatsClicksParams {
     /// The number of periods to include in the response.
     #[uniffi(default = None)]
     pub num: Option<u32>,
-    /// The locale for the response.
-    #[uniffi(default = None)]
-    pub locale: Option<WPComLanguage>,
     /// Whether to return a summary of the data.
     ///
     /// - `true` (default): Response contains `summary` field with aggregated data
@@ -78,7 +74,6 @@ impl Default for StatsClicksParams {
             start_date: None,
             max: None,
             num: None,
-            locale: None,
             summarize: true,
             skip_archives: true,
         }
@@ -93,7 +88,6 @@ impl AppendUrlQueryPairs for StatsClicksParams {
             .append_option_query_value_pair("start_date", self.start_date.as_ref())
             .append_option_query_value_pair("max", self.max.as_ref())
             .append_option_query_value_pair("num", self.num.as_ref())
-            .append_option_query_value_pair("locale", self.locale.as_ref())
             .append_query_value_pair("summarize", &(self.summarize as u32))
             .append_query_value_pair("skip_archives", &(self.skip_archives as u32));
     }
@@ -181,7 +175,6 @@ mod tests {
             start_date: Some(WpDateString::new("2026-02-18".to_string())),
             max: Some(10),
             num: Some(30),
-            locale: Some(WPComLanguage::English),
             summarize: true,
             skip_archives: true,
         };
@@ -191,7 +184,7 @@ mod tests {
 
         assert_eq!(
             query_pairs.finish().as_str(),
-            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/clicks?period=day&date=2026-02-18&start_date=2026-02-18&max=10&num=30&locale=en&summarize=1&skip_archives=1"
+            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/clicks?period=day&date=2026-02-18&start_date=2026-02-18&max=10&num=30&summarize=1&skip_archives=1"
         );
     }
 
@@ -207,7 +200,6 @@ mod tests {
             start_date: None,
             max: None,
             num: None,
-            locale: None,
             summarize: true,
             skip_archives: true,
         };