+++ wp_api/src/wp_com/stats_referrers.rs
@@ -2,7 +2,6 @@ use crate::{
     date::WpDateString,
     impl_as_query_value_from_to_string,
     url_query::{AppendUrlQueryPairs, QueryPairs, QueryPairsExtension},
-    wp_com::language::WPComLanguage,
 };
 use serde::{Deserialize, Deserializer, Serialize};
 use std::collections::HashMap;
@@ -53,9 +52,6 @@ pub struct StatsReferrersParams {
     /// The number of periods to include in the response.
     #[uniffi(default = None)]
     pub num: Option<u32>,
-    /// The locale for the response.
-    #[uniffi(default = None)]
-    pub locale: Option<WPComLanguage>,
     /// Whether to return a summary of the data.
     ///
     /// - `true` (default): Response contains `summary` field with aggregated data
@@ -79,7 +75,6 @@ impl Default for StatsReferrersParams {
             start_date: None,
             max: None,
             num: None,
-            locale: None,
             summarize: true,
             skip_archives: Some(true),
         }
@@ -94,7 +89,6 @@ impl AppendUrlQueryPairs for StatsReferrersParams {
             .append_option_query_value_pair("start_date", self.start_date.as_ref())
             .append_option_query_value_pair("max", self.max.as_ref())
             .append_option_query_value_pair("num", self.num.as_ref())
-            .append_option_query_value_pair("locale", self.locale.as_ref())
             .append_query_value_pair("summarize", &(self.summarize as u32))
             .append_option_query_value_pair(
                 "skip_archives",
@@ -274,7 +268,6 @@ mod tests {
             start_date: Some(WpDateString::new("2026-01-26".to_string())),
             max: Some(10),
             num: Some(30),
-            locale: Some(WPComLanguage::English),
             summarize: true,
             skip_archives: Some(true),
         };
@@ -284,7 +277,7 @@ mod tests {
 
         assert_eq!(
             query_pairs.finish().as_str(),
-            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/referrers?period=day&date=2026-01-26&start_date=2026-01-26&max=10&num=30&locale=en&summarize=1&skip_archives=1"
+            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/referrers?period=day&date=2026-01-26&start_date=2026-01-26&max=10&num=30&summarize=1&skip_archives=1"
         );
     }
 
@@ -301,7 +294,6 @@ mod tests {
             start_date: None,
             max: None,
             num: None,
-            locale: None,
             summarize: true,
             skip_archives: Some(true),
         };