+++ wp_api/src/wp_com/stats_tags.rs
@@ -1,7 +1,6 @@
 use crate::{
     date::WpDateString,
     url_query::{AppendUrlQueryPairs, QueryPairs, QueryPairsExtension},
-    wp_com::language::WPComLanguage,
 };
 use serde::{Deserialize, Serialize};
 
@@ -11,16 +10,11 @@ pub struct StatsTagsParams {
     /// The maximum number of tags to return.
     #[uniffi(default = None)]
     pub max: Option<u32>,
-    /// The locale for the response.
-    #[uniffi(default = None)]
-    pub locale: Option<WPComLanguage>,
 }
 
 impl AppendUrlQueryPairs for StatsTagsParams {
     fn append_query_pairs(&self, query_pairs_mut: &mut QueryPairs) {
-        query_pairs_mut
-            .append_option_query_value_pair("max", self.max.as_ref())
-            .append_option_query_value_pair("locale", self.locale.as_ref());
+        query_pairs_mut.append_option_query_value_pair("max", self.max.as_ref());
     }
 }
 
@@ -64,17 +58,14 @@ mod tests {
             url::Url::parse("https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/tags")
                 .expect("Failed to parse url");
 
-        let params = StatsTagsParams {
-            max: Some(7),
-            locale: Some(WPComLanguage::English),
-        };
+        let params = StatsTagsParams { max: Some(7) };
 
         let mut query_pairs = url.query_pairs_mut();
         params.append_query_pairs(&mut query_pairs);
 
         assert_eq!(
             query_pairs.finish().as_str(),
-            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/tags?max=7&locale=en"
+            "https://public-api.wordpress.com/rest/v1.1/sites/1234/stats/tags?max=7"
         );
     }